Split-up DeepLC class into separate functions (WIP)
